feat: periodic cleanup of expired chunks from the reserve
